requirements.txt
albucore==0.0.24                                                                                                                                                                                                                                        
albumentations==2.0.8
iterative-stratification==0.1.9
numpy==2.3.2
opencv-python==4.11.0.86
pandas==2.3.1
Pillow==9.5.0
PyYAML==6.0.2
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.16.1
timm==1.0.19
tqdm==4.67.1
pydicom==3.0.1
gdcm==1.1
pylibjpeg==2.0.1 
pylibjpeg-libjpeg==2.2.0
simpleitk==2.5.2
ultralytics==8.3.184
ipython==9.5.0
seaborn==0.13.2
tensorboard==2.20.0
nibabel==5.3.2